requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
Pillow==12.1.1
python-decouple==3.8
asgiref==3.11.1
sqlparse==0.5.5
PyJWT==2.7.0
